refactoring the codes
