+++ offworld_gym/envs/gazebo/__init__.py
@@ -1,7 +1,4 @@
 """
-This module contains Gazebo-based OffWorld Gym environments
+This module contains Dockerized Gazebo-based OffWorld Gym environments
 """
-
-from offworld_gym.envs.gazebo.gazebo_env import GazeboGymEnv
-# from offworld_gym.envs.gazebo.offworld_monolith_env import OffWorldMonolithDiscreteEnv, OffWorldMonolithContinuousEnv
-from offworld_gym.envs.gazebo.offworld_monolith_obstacle_env import OffWorldMonolithObstacleDiscreteEnv, OffWorldMonolithObstacleContinuousEnv
+from offworld_gym.envs.gazebo.offworld_monolith_env import OffWorldDockerMonolithDiscreteEnv, OffWorldDockerMonolithContinuousEnv
