+++ offworld_gym/envs/gazebo/docker_entrypoint.sh
