+++ offworld_gym/envs/gazebo/gazebo_utils.py
@@ -20,17 +20,16 @@ import numpy as np
 import cv2
 import io
 from imageio import imread
+import base64
+import time
 import rospy
 from std_srvs.srv import Empty as Empty_srv
 from gazebo_msgs.srv import GetModelState, SetModelState
 from gazebo_msgs.msg import ModelState
-import base64
-import time
 
 class GazeboUtils:
     """Gazebo utility functions used by the OffWorld Gym environments
     """
-
     unpause = rospy.ServiceProxy('/gazebo/unpause_physics', Empty_srv)
     pause = rospy.ServiceProxy('/gazebo/pause_physics', Empty_srv)
 
@@ -72,50 +71,4 @@ class GazeboUtils:
             rospy.logerr('The robot cannot be reset.')
             raise e
 
-
-class ImageUtils(object):
-    """Image utility functions used by the OffWorld Gym environments
-    """
-    IMG_H = 240
-    IMG_W = 320
-    IMG_C = 3
-
-    @staticmethod
-    def process_img_msg(img_msg, resized_width=IMG_W, resized_height=IMG_H, max_value_for_clip_and_normalize=None):
-        """Converts ROS image to cv2, then to numpy
-        """
-        base64_bytes = img_msg['data'].encode('ascii')
-        image_bytes = base64.b64decode(base64_bytes)
-        np_image = np.frombuffer(image_bytes, dtype=np.uint8)
-        np_image = np_image.reshape((img_msg['height'],img_msg['width'],-1))
-        img = np_image[:, :, ::-1].copy() # replace cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-        # img = cv2.resize(img, (resized_width, resized_height)) # could cause img corrupted
-        img = np.reshape(img, (1, img.shape[0], img.shape[1], img.shape[2]))
-
-        if max_value_for_clip_and_normalize is not None:
-            img = np.clip(img, a_min=0.0, a_max=max_value_for_clip_and_normalize)
-            img = img / max_value_for_clip_and_normalize
- 
-        return img 
-    
-    @staticmethod
-    def process_depth_msg(depth_msg, resized_width=IMG_W, resized_height=IMG_H, max_value_for_clip_and_normalize=None):
-        """Converts a depth image into numpy float32 array
-        """
-
-        base64_bytes = depth_msg['data'].encode('ascii')
-        image_bytes = base64.b64decode(base64_bytes)
-        np_image = np.frombuffer(image_bytes, dtype=np.float32)
-        np_image = np_image.reshape((depth_msg['height'],depth_msg['width'],-1))
-        # img = depth_msg.astype("float32")
-        img = np.nan_to_num(np_image)
-        # img = cv2.resize(img, (resized_width, resized_height)) # could cause img corrupted
-
-        img = np.reshape(img, (1, img.shape[0], img.shape[1], 1))
-
-        if max_value_for_clip_and_normalize is not None:
-            img = np.clip(img, a_min=0.0, a_max=max_value_for_clip_and_normalize)
-            img = img / max_value_for_clip_and_normalize
-
-        return img
