+++ offworld_gym/envs/gazebo/generate_protobuf_code.sh
