+++ offworld_gym/envs/gazebo/offworld_monolith_env.py
@@ -17,41 +17,38 @@ from offworld_gym import version
 __version__     = version.__version__
 
 #lib
+import ast
+import os
 from math import pi
 import time
+import subprocess
 import numpy as np
 from scipy.spatial import distance
-import pdb
+import json
+import requests
+
 from pyquaternion import Quaternion
 from matplotlib import pyplot as plt
-
-
+from typing import List, Tuple, Union, Any
 #gym
 import gym
 from gym import utils, spaces
-from offworld_gym.envs.gazebo.gazebo_env import GazeboGymEnv
+from offworld_gym.envs.gazebo.docker_gazebo_env import DockerizedGazeboEnv
 from gym.utils import seeding
-from offworld_gym.envs.gazebo.utils import ImageUtils, GazeboUtils
+from offworld_gym.envs.gazebo.image_utils import ImageUtils
 from offworld_gym.envs.common.exception.gym_exception import GymException
 from offworld_gym.envs.common.channels import Channels
 from offworld_gym.envs.common.actions import FourDiscreteMotionActions
 
 #ros
-import rospy
-from std_srvs.srv import Empty as Empty_srv
-from geometry_msgs.msg import Twist
-from nav_msgs.msg import Odometry
-from gazebo_msgs.srv import GetModelState
-from gazebo_msgs.msg import ModelState
-from sensor_msgs.msg import Image
-from rosgraph_msgs.msg import Clock
+import roslibpy
 
-# import logging
-# logger = logging.getLogger(__name__)
-# level = logging.DEBUG
-# logger.setLevel(level)
+import logging
+logger = logging.getLogger(__name__)
+level = logging.DEBUG
+logger.setLevel(level)
 
-class OffWorldMonolithEnv(GazeboGymEnv):
+class OffWorldDockerizedMonolithEnv(DockerizedGazeboEnv):
     """Generic Simulated gym environment that replicates the real OffWorld Monolith environment in Gazebo.
 
     Agent receives RGB or Depth camera feed as input and needs to learn to approch the visual
@@ -69,45 +66,35 @@ class OffWorldMonolithEnv(GazeboGymEnv):
     _EPISODE_LENGTH = 100
     _STEP_DURATION_SECONDS_IN_SIM = 1.0
     _MAX_TOLERABLE_ROSLAUNCH_INIT_SECONDS = 20
-    _WALL_BOUNDARIES = {"x_max": 1.90, "x_min": -1.75, "y_max": 1.10, "y_min": -1.60}
+    _WALL_BOUNDARIES = {"x_max": 1.80, "x_min": -1.60, "y_max": 1.00, "y_min": -1.45}
 
     def __init__(self, channel_type=Channels.DEPTH_ONLY, random_init=True):
 
-        super(OffWorldMonolithEnv, self).__init__(package_name='gym_offworld_monolith', launch_file='env_bringup.launch')
+        super(OffWorldDockerizedMonolithEnv, self).__init__(package_name='gym_offworld_monolith', launch_file='env_bringup.launch')
 
         assert isinstance(channel_type, Channels), "Channel type is not of Channels."
 
         before_ros_init = time.time()
 
         # Avoid race condition with rosbot launch by waiting for it fully init and start publishing it's odometry
-        self._odom_called = False
-        def odom_call(msg):
-            self._odom_called = True
-        GazeboUtils.unpause_physics()
+        self.unpause_sim = self.register_ros_service('/gazebo/unpause_physics', "std_srvs/Empty_srv")
+        self.unpause_physics_remotely() # rosbridge replaced by endpoint python server
         print("waiting for rosbot model to init...")
-        self.vel_sub = rospy.Subscriber('odom', Odometry, odom_call, queue_size=1)
-        while not self._odom_called:
+        self.vel_sub = self.register_subscriber('/odom', "nav_msgs/Odometry", self._latest_odom_message, queue_size=1)
+        
+        while not self.vel_sub._subscribe_id:
             if time.time() - before_ros_init > self._MAX_TOLERABLE_ROSLAUNCH_INIT_SECONDS:
                 raise GymException("ROS took too long to for the rosbot diff_drive_controller to publish to /odom")
             time.sleep(0.1)
 
-        # be certain that /clock is being published, so we can use it to track sim-time
+        # Be certain that /clock is being published, so we can use it to track sim-time
         print("waiting for publications to /clock...")
-        self._latest_clock_message: Clock = None
-        def update_sim_time_from_clock(msg: Clock):
-            self._latest_clock_message = msg
-        self.clock_sub = rospy.Subscriber('/clock', Clock, update_sim_time_from_clock, queue_size=1)
-
-        self._latest_depth_img = None
-        def update_depth_img(msg):
-            self._latest_depth_img = msg
-        self.depth_sub = rospy.Subscriber('/camera/depth/image_raw', Image, update_depth_img, queue_size=1)
-
-        self._latest_rgb_img = None
-        def update_rgb_img(msg):
-            self._latest_rgb_img = msg
-        self.depth_sub = rospy.Subscriber('/camera/rgb/image_raw', Image, update_rgb_img, queue_size=1)
-
+        self.clock_sub = self.register_subscriber('/clock', "rosgraph_msgs/Clock", self._latest_clock_message, queue_size=1)
+    
+        # Subscribe to img topics
+        self.depth_sub = self.register_subscriber('/camera/depth/image_raw', "sensor_msgs/Image", self._latest_depth_img, queue_size=1)
+        self.rbg_sub = self.register_subscriber('/camera/rgb/image_raw', "sensor_msgs/Image", self._latest_rgb_img, queue_size=1)
+        
         while self._latest_clock_message is None:
             if time.time() - before_ros_init > self._MAX_TOLERABLE_ROSLAUNCH_INIT_SECONDS:
                 raise GymException("ROS took too long to publish to /clock")
@@ -123,15 +110,18 @@ class OffWorldMonolithEnv(GazeboGymEnv):
                 raise GymException("ROS took too long to publish to /camera/rgb/image_raw")
             time.sleep(0.1)
 
-        GazeboUtils.pause_physics()
-        rospy.loginfo("Environment has been initiated.")
+        self.pause_sim = self.register_ros_service('/gazebo/pause_physics', "std_srvs/Empty_srv")
+        self.pause_physics_remotely() # rosbridge replaced by endpoint python server
+        logger.info("Environment has been initiated.")
 
         # gazebo
-        self.reset_sim = rospy.ServiceProxy('/gazebo/reset_simulation', Empty_srv)
-        rospy.loginfo("Service proxies have been initiated.")
+        self.reset_sim = self.register_ros_service('/gazebo/reset_simulation', "std_srvs/Empty_srv")
+        self.reset_sim_respond = self.call_ros_service(self.reset_sim,'/gazebo/reset_simulation', "std_srvs/Empty_srv")
+        logger.info("Service proxies have been initiated.")
 
-        # rosbot
-        self.vel_pub = rospy.Publisher('/cmd_vel', Twist, queue_size=2)
+        # initialize other service proxy
+        self.get_model_state_srv = self.register_ros_service( "/gazebo/get_model_state", "gazebo_msgs/GetModelState")
+        self.set_robot_state_srv = self.register_ros_service( "/gazebo/set_model_state", "gazebo_msgs/SetModelState")
 
         # environment
         self.seed()
@@ -144,18 +134,18 @@ class OffWorldMonolithEnv(GazeboGymEnv):
         self.observation_space = spaces.Box(0, 255, shape = (1, ImageUtils.IMG_H, ImageUtils.IMG_W, channel_type.value))
         self.action_space = None
         self._monolith_space = self._get_state_vector('monolith')
-        rospy.logdebug("----------------Monolith----------------")
-        rospy.logdebug("Pose x: {}".format(str(self._monolith_space[0])))
-        rospy.logdebug("Pose y: {}".format(str(self._monolith_space[1])))
-        rospy.logdebug("Pose z: {}".format(str(self._monolith_space[2])))
-        rospy.logdebug("----------------------------------------")
-        rospy.loginfo("Environment has been started.")
+        logger.debug("----------------Monolith----------------")
+        logger.debug("Pose x: {}".format(str(self._monolith_space[0])))
+        logger.debug("Pose y: {}".format(str(self._monolith_space[1])))
+        logger.debug("Pose z: {}".format(str(self._monolith_space[2])))
+        logger.debug("----------------------------------------")
+        logger.info("Environment has been started.")
 
     def _sim_time_sleep(self, sleep_seconds):
-        before_sleep_sim_time_secs = self._latest_clock_message.clock.secs + self._latest_clock_message.clock.nsecs/1e+9
+        before_sleep_sim_time_secs = self._latest_clock_message["clock"]["secs"] + self._latest_clock_message["clock"]["nsecs"]/1e+9
         while True:
             time.sleep(0.0001)
-            cur_sim_time_sec = self._latest_clock_message.clock.secs + self._latest_clock_message.clock.nsecs/1e+9
+            cur_sim_time_sec = self._latest_clock_message["clock"]["secs"] + self._latest_clock_message["clock"]["nsecs"]/1e+9
             if cur_sim_time_sec - before_sleep_sim_time_secs > sleep_seconds:
                 break
 
@@ -174,6 +164,7 @@ class OffWorldMonolithEnv(GazeboGymEnv):
         Returns:
             Numpy array with the state of the environment as captured by the robot's rgbd sensor.
         """
+        # import pdb; pdb.set_trace()
         rgb_img = ImageUtils.process_img_msg(self._latest_rgb_img)
         depth_img = ImageUtils.process_depth_msg(self._latest_depth_img)
 
@@ -183,10 +174,10 @@ class OffWorldMonolithEnv(GazeboGymEnv):
             state = rgb_img
         elif self.channel_type == Channels.RGBD:
             state = np.concatenate((rgb_img, depth_img), axis=-1)
-        rospy.loginfo("State of the environment captured.")
+        logger.debug("State of the environment captured.")
         return state
 
-    def _move_rosbot(self, lin_x_speed, ang_z_speed, sleep_time=2.):
+    def _move_rosbot(self, lin_x_speed, ang_z_speed, sleep_time=2.0):
         """Moves the ROSBot.
 
         Accepts linear x speed and angular z speed and moves the
@@ -199,19 +190,20 @@ class OffWorldMonolithEnv(GazeboGymEnv):
         Returns:
             The real time factor for the move (sim-time elapsed/wall-time elapsed)
         """
-        vel_cmd = Twist()
-        vel_cmd.linear.x = lin_x_speed
-        vel_cmd.angular.z = ang_z_speed
-        self.vel_pub.publish(vel_cmd)
-
-        GazeboUtils.unpause_physics()
+        self.unpause_physics_remotely() # rosbridge replaced by endpont python server
         wall_start = time.time()
+        if self._rosbridge_client.is_connected:
+            self.publish_cmd_vel_remotely(lin_x_speed, ang_z_speed)  # rosbridge replaced by endpont python server
+
         self._sim_time_sleep(sleep_time)  # action duration is in sim-time, so simulation speed has no affect on env dynamics
         wall_stop = time.time()
-        GazeboUtils.pause_physics()
-
+        self.pause_physics_remotely() # rosbridge replaced by endpont python server
+        
         wall_sleep_time = wall_stop - wall_start
+        logger.debug("Profilling action and pause service {}".format(str(wall_sleep_time)))
         real_time_factor = sleep_time / wall_sleep_time
+        logger.debug("real time factor : {} ".format(str(real_time_factor)))
+    
         return real_time_factor
 
     def _get_model_state(self, name):
@@ -224,12 +216,13 @@ class OffWorldMonolithEnv(GazeboGymEnv):
             A vector containing state of the model contained in a GetModelState service message.
         """
         try:
-            g_getStateVector = rospy.ServiceProxy("/gazebo/get_model_state", GetModelState)
-            rospy.wait_for_service("/gazebo/get_model_state")
-            state = g_getStateVector(model_name=name)
+            # rosbridge returns a dict(service responce obj)
+            state = self.call_ros_service(self.get_model_state_srv, "/gazebo/get_model_state", "gazebo_msgs/GetModelState", {"model_name": f"{name}"})
+            logger.debug("Current robot state {}".format(str(state['pose'])))
             return state
-        except rospy.ServiceException:
-            rospy.logerr("An error occured while invoking the get_model_state service.")
+            
+        except Exception as e:
+            logger.error("An error occured while invoking the get_model_state service.")
             return None
 
     def _get_state_vector(self, name):
@@ -244,14 +237,14 @@ class OffWorldMonolithEnv(GazeboGymEnv):
         try:
             ms = self._get_model_state(name)
             if ms is not None:
-                quaternion = Quaternion(ms.pose.orientation.w,  ms.pose.orientation.x, ms.pose.orientation.y, ms.pose.orientation.z,)
+                quaternion = Quaternion(ms['pose']['orientation']['w'], ms['pose']['orientation']['x'], ms['pose']['orientation']['y'], ms['pose']['orientation']['z'])
                 yaw, _, _ = quaternion.yaw_pitch_roll
-                state = (ms.pose.position.x, ms.pose.position.y, ms.pose.position.z, yaw)
+                state = (ms['pose']['position']['x'], ms['pose']['position']['y'], ms['pose']['position']['z'], yaw)
                 return state
             else:
                 return None
-        except rospy.ServiceException:
-            rospy.logerr("An error occured while creating the state vector.")
+        except Exception as e:
+            logger.error("An error occured while creating the state vector.")
             return None
 
     def _calculate_reward(self, action):
@@ -263,13 +256,13 @@ class OffWorldMonolithEnv(GazeboGymEnv):
         """
         rosbot_state = self._get_state_vector('rosbot')
         dst = distance.euclidean(rosbot_state[0:3], self._monolith_space[0:3])
-        rospy.logdebug("Distance between the rosbot and the monolith is : {}".format(str(dst)))
+        logger.debug("Distance between the rosbot and the monolith is : {}".format(str(dst)))
 
         # check distance to the monolith
 
         # for discrete action space
         if self.action_space == spaces.Discrete(4): 
-            if dst < OffWorldMonolithEnv._PROXIMITY_THRESHOLD:
+            if dst < OffWorldDockerizedMonolithEnv._PROXIMITY_THRESHOLD:
                 if self._backward_step_counter == 0 and action == FourDiscreteMotionActions.BACKWARD:
                     reward = 0.0 # give another chance, not ending the episode yet
                     done = False 
@@ -288,7 +281,7 @@ class OffWorldMonolithEnv(GazeboGymEnv):
                
         # for continuous action space
         else: 
-            if dst < OffWorldMonolithEnv._PROXIMITY_THRESHOLD:
+            if dst < OffWorldDockerizedMonolithEnv._PROXIMITY_THRESHOLD:
                 if self._backward_step_counter == 0 and action[0] <= 0.0:
                     reward = 0.0 # give another chance, not ending the episode yet
                     done = False 
@@ -305,8 +298,6 @@ class OffWorldMonolithEnv(GazeboGymEnv):
                 reward = 0.0 
                 done = False 
 
-        rospy.loginfo(f"\n New reward logic")
-
         # check boundaries
         if rosbot_state[0] < self._WALL_BOUNDARIES['x_min'] or \
                 rosbot_state[0] > self._WALL_BOUNDARIES['x_max'] or \
@@ -315,8 +306,10 @@ class OffWorldMonolithEnv(GazeboGymEnv):
             reward = 0.0
             done = True
 
-        if self.step_count == OffWorldMonolithEnv._EPISODE_LENGTH:
+        if self.step_count == OffWorldDockerizedMonolithEnv._EPISODE_LENGTH:
             done = True
+
+        logger.debug("Done : {}".format(bool(done)))
         return reward, done
 
     def step(self, action):
@@ -334,26 +327,34 @@ class OffWorldMonolithEnv(GazeboGymEnv):
         """
         try:
             assert model_name is not None or model_name != ''
-            goal_state = ModelState()
-            goal_state.model_name = model_name
+            goal_state_model_name = model_name
 
             # random spawn location exlucing no-spawn 30cm radius around the monolith
-            goal_state.pose.position.x = self._monolith_space[0]
-            goal_state.pose.position.y = self._monolith_space[1]
-            while distance.euclidean((goal_state.pose.position.x, goal_state.pose.position.y), self._monolith_space[0:2]) < 0.50:
-                goal_state.pose.position.x = np.random.uniform(low=self._WALL_BOUNDARIES['x_min'] + 0.08,
+            goal_state_pose_position_x = self._monolith_space[0]
+            goal_state_pose_position_y = self._monolith_space[1]
+            while distance.euclidean((goal_state_pose_position_x, goal_state_pose_position_y), self._monolith_space[0:2]) < 0.50:
+                goal_state_pose_position_x = np.random.uniform(low=self._WALL_BOUNDARIES['x_min'] + 0.08,
                                                                high=self._WALL_BOUNDARIES['x_max'] - 0.08)
-                goal_state.pose.position.y = np.random.uniform(low=self._WALL_BOUNDARIES['y_min'] + 0.08,
+                goal_state_pose_position_y = np.random.uniform(low=self._WALL_BOUNDARIES['y_min'] + 0.08,
                                                                high=self._WALL_BOUNDARIES['y_max'] - 0.08)
 
-            rospy.loginfo("Spawning at (%.2f, %.2f)" % (goal_state.pose.position.x, goal_state.pose.position.y))
-            goal_state.pose.position.z = 0.20
+            logger.info("Spawning at (%.2f, %.2f)" % (goal_state_pose_position_x, goal_state_pose_position_y))
+            goal_state_pose_position_z = 0.20
 
             if orient:
-                goal_state.pose.orientation.z, goal_state.pose.orientation.w = np.random.uniform(-3.14, 3.14, size=(2,))
-            GazeboUtils.move_to_position(goal_state)
-        except rospy.ROSException:
-            rospy.logerr("An error occured while resetting the environment.")
+                goal_state_pose_orientation_z, goal_state_pose_orientation_w = np.random.uniform(-3.14, 3.14, size=(2,))
+            
+            goal_state_dict = {"model_state": 
+                            {"model_name":goal_state_model_name, 
+                            "pose":
+                            {"position" : {"x" : goal_state_pose_position_x, "y" : goal_state_pose_position_y, "z" : goal_state_pose_position_z} ,
+                            "orientation":{"x": 0.0, "y": 0.0, "z": goal_state_pose_orientation_z, "w": goal_state_pose_orientation_w}}}}
+
+
+            move_to_position = self.call_ros_service(self.set_robot_state_srv, "/gazebo/set_model_state", "gazebo_msgs/SetModelState", goal_state_dict)
+       
+        except Exception as e:
+            logger.error("An error occured while resetting the environment.")
             raise GymException("An error occured while resetting the environment.")
 
     def _move_to_original_position(self, model_name):
@@ -368,15 +369,21 @@ class OffWorldMonolithEnv(GazeboGymEnv):
         try:
             assert model_name is not None or model_name != ''
 
-            goal_state = ModelState()
-            goal_state.model_name = model_name
-            goal_state.pose.position.x = 0.0
-            goal_state.pose.position.y = 0.0
-            goal_state.pose.position.z = 0.20
+            # goal_state = ModelState()
+            goal_state_model_name = model_name
+            goal_state_pose_position_x = 0.0
+            goal_state_pose_position_y = 0.0
+            goal_state_pose_position_z = 0.20
 
-            GazeboUtils.move_to_position(goal_state)
-        except rospy.ServiceException:
-            rospy.logerr("An error occured while resetting the environment.")
+            goal_state_dict = {"model_state": 
+                            {"model_name":goal_state_model_name, 
+                            "pose":
+                            {"position" : {"x" : goal_state_pose_position_x, "y" : goal_state_pose_position_y, "z" : goal_state_pose_position_z}}}}
+
+
+            move_to_position = self.call_ros_service(self.set_robot_state_srv, "/gazebo/set_model_state", "gazebo_msgs/SetModelState", goal_state_dict)
+        except Exception as e:
+            logger.error("An error occured while resetting the environment.")
             raise GymException("An error occured while resetting the environment.")
 
     def reset(self):
@@ -386,18 +393,16 @@ class OffWorldMonolithEnv(GazeboGymEnv):
             A numpy array with rgb/depth/rgbd encoding of the state as seen by
             the robot
         """
-        GazeboUtils.pause_physics()
-        vel_cmd = Twist()
-        vel_cmd.linear.x = 0.0
-        vel_cmd.angular.z = 0.0
-        self.vel_pub.publish(vel_cmd)
+        self.pause_physics_remotely()
+        if self._rosbridge_client.is_connected:
+            self.publish_cmd_vel_remotely(lin_x_speed=0.0, ang_z_speed=0.0)
 
         if self.random_init:
             self._move_to_random_position('rosbot')
         else:
             self._move_to_original_position('rosbot')
-
         state = self._get_state()
+        self.unpause_physics_remotely()
         return state
 
     def render(self, mode='human'):
@@ -438,7 +443,7 @@ class OffWorldMonolithEnv(GazeboGymEnv):
             plt.show(block=False)
             plt.pause(0.05)
 
-class OffWorldMonolithDiscreteEnv(OffWorldMonolithEnv):
+class OffWorldDockerMonolithDiscreteEnv(OffWorldDockerizedMonolithEnv):
     """Discrete version of the simulated gym environment that replicates the real OffWorld Monolith environment in Gazebo.
 
     .. code:: python
@@ -449,7 +454,7 @@ class OffWorldMonolithDiscreteEnv(OffWorldMonolithEnv):
     """
 
     def __init__(self, channel_type=Channels.DEPTH_ONLY, random_init=True):
-        super(OffWorldMonolithDiscreteEnv, self).__init__(channel_type=channel_type, random_init=random_init)
+        super(OffWorldDockerMonolithDiscreteEnv, self).__init__(channel_type=channel_type, random_init=random_init)
         self.action_space = spaces.Discrete(4)
 
     def _send_action_commands(self, action_type):
@@ -493,8 +498,8 @@ class OffWorldMonolithDiscreteEnv(OffWorldMonolithEnv):
             assert action >= 0 and action < 4, "Unrecognized value for the action"
             action = FourDiscreteMotionActions(action)
 
-        rospy.loginfo("Step: %d" % self.step_count)
-        rospy.loginfo(action)
+        logger.info("Step: %d" % self.step_count)
+        logger.info(action)
         real_time_factor_for_move = self._send_action_commands(action)
 
         self._current_state = self._get_state()
@@ -507,7 +512,7 @@ class OffWorldMonolithDiscreteEnv(OffWorldMonolithEnv):
         return self._current_state, reward, done, info
 
 
-class OffWorldMonolithContinuousEnv(OffWorldMonolithEnv):
+class OffWorldDockerMonolithContinuousEnv(OffWorldDockerizedMonolithEnv):
     """Continous version of the simulated gym environment that replicates the real OffWorld Monolith environment in Gazebo.
 
     .. code:: python
@@ -518,7 +523,7 @@ class OffWorldMonolithContinuousEnv(OffWorldMonolithEnv):
     """
 
     def __init__(self, channel_type=Channels.DEPTH_ONLY, random_init=True):
-        super(OffWorldMonolithContinuousEnv, self).__init__(channel_type=channel_type, random_init=random_init)
+        super(OffWorldDockerMonolithContinuousEnv, self).__init__(channel_type=channel_type, random_init=random_init)
         self.action_space = spaces.Box(low=np.array([-0.7, -2.5]), high=np.array([0.7, 2.5]), dtype=np.float32)
         self.action_limit = np.array([[-0.7, -2.5], [0.7, 2.5]])
 
@@ -539,8 +544,8 @@ class OffWorldMonolithContinuousEnv(OffWorldMonolithEnv):
         assert action is not None, "Action cannot be None."
         assert isinstance(action, (np.ndarray)), "Action type is not recognized."
         action = np.clip(action, self.action_limit[0], self.action_limit[1])
-        rospy.loginfo("Step: %d" % self.step_count)
-        rospy.loginfo(action)
+        logger.info("Step: %d" % self.step_count)
+        logger.info(action)
         real_time_factor_for_move = self._move_rosbot(action[0], action[1], self._STEP_DURATION_SECONDS_IN_SIM)
 
         self._current_state = self._get_state()
