--- offworld_gym/envs/gazebo_docker/__init__.py
@@ -1,4 +1,0 @@
-"""
-This module contains Dockerized Gazebo-based OffWorld Gym environments
-"""
-from offworld_gym.envs.gazebo_docker.offworld_monolith_env import OffWorldDockerMonolithDiscreteEnv, OffWorldDockerMonolithContinuousEnv
